blogdata/requirements.txt
openai==1.35.7
python-dotenv==1.0.1
playwright==1.44.0
beautifulsoup4==4.12.3
requests==2.32.3
tqdm==4.66.4
httpx==0.27.0
pytz==2024.1
aiohttp==3.9.5
aiofiles==23.2.1
boto3==1.34.131
aioboto3==12.3.0
